Initial commit.
This is the first commit of the project. Here the base of the project
has been created with all vital files needed to create and run the
project. The app will now display the root widget which contains the
title of the app.
In future, unnecessary files will be removed and possibly more vital
files added.