Raw list is displayed.
This is a very early stage list of where the forcast is to be displayed.
Widgets have been set up in different folders, Changing of state needs
to be handled, which is in the next steps. Before that can happen, an
input field needs to be created for the user to enter a location which
will then impact the state.