Button and textfield have been added.
A textfield and a button below have been added to the UI and are
connected. This is the final setup of the application and the first
version of the app. It has a raw structure to build on and a good
checkpoint for future development, until next version is pushed to the
master branch.