Data fetched from API and displayd(very raw).
Two inputfields can now be used to fetch weather forcast about the next
five days. The weather can be eather fetched by city name or
coordinates and is displayed below in a raw form.
The look of the whole thing is very raw and will be fixed. Next steps
should be displaying the data in an efficient way and making items
clickable to see further info about future days.