Weather displayed for the next 5 days with detail.
Weather is now displayd for the next five days with the possibility to
press each day for more detail. The UI has been updated to meet the
commands, to have the correct widgets and in the right order to make
this happen. The next step is to display only the first day and the rest
below.