Caching of last search input is complete.
The last weather search (eather by city or coordinates) is cached for a
maximum of twelve hours. A decision was made to store the last search
instead of the last data displayed since the weather information will
run out of date when time passes, so that it would be more useful for
the user to see the latest forcast for the last search.
Next steps are to let the user have the option to clear the cache.`